+++ bindings/python/pinocchio/cppad/__init__.py
@@ -26,3 +26,6 @@ if WITH_COLLISION:
         StdVec_Contact,
         StdVec_DistanceResult,
     )
+
+    # Deprecated, should be removed in next major release
+    hppfcl = coal