+++ include/pinocchio/bindings/python/collision/fcl/transform.hpp
@@ -1,0 +1,16 @@
+//
+// Copyright (c) 2026 INRIA
+//
+
+#ifndef __pinocchio_python_collision_fcl_transform_hpp__
+#define __pinocchio_python_collision_fcl_transform_hpp__
+
+#include "pinocchio/macros.hpp"
+
+// clang-format off
+PINOCCHIO_MOVED_HEADER_PINOCCHIO4(pinocchio/bindings/python/collision/fcl/transform.hpp, pinocchio/bindings/python/collision/coal/transform.hpp)
+// clang-format on
+
+#include "pinocchio/bindings/python/collision/coal/transform.hpp"
+
+#endif // ifndef __pinocchio_python_collision_fcl_transform_hpp__