+++ include/pinocchio/serialization/fcl.hpp
@@ -1,0 +1,16 @@
+//
+// Copyright (c) 2026 INRIA
+//
+
+#ifndef __pinocchio_multibody_fcl_serialization_hpp__
+#define __pinocchio_multibody_fcl_serialization_hpp__
+
+#include "pinocchio/macros.hpp"
+
+// clang-format off
+PINOCCHIO_MOVED_HEADER_PINOCCHIO4(pinocchio/serialization/fcl.hpp, pinocchio/serialization/coal.hpp)
+// clang-format on
+
+#include "pinocchio/serialization/coal.hpp"
+
+#endif // __pinocchio_multibody_fcl_serialization_hpp__